requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.2
iso8601~=0.1.16
rfc3339~=6.2